+++ src/block_sync.h
@@ -10,10 +10,6 @@
 
 namespace redsea {
 
-enum eOffset {
-  OFFSET_A, OFFSET_B, OFFSET_C, OFFSET_CI, OFFSET_D, OFFSET_INVALID
-};
-
 enum eInputType {
   INPUT_MPX, INPUT_ASCIIBITS, INPUT_RDSSPY
 };
@@ -47,15 +43,12 @@ class BlockStream {
   eOffset received_offset_;
   uint16_t pi_;
   bool is_in_sync_;
-  std::vector<uint16_t> group_data_;
-  std::vector<bool> has_block_;
   std::vector<bool> block_has_errors_;
 #ifdef HAVE_LIQUID
   Subcarrier subcarrier_;
 #endif
   AsciiBits ascii_bits_;
   std::map<uint16_t,uint32_t> error_lookup_;
-  unsigned num_blocks_received_;
   const eInputType input_type_;
   bool is_eof_;